refactor(kids-info): Moving all Kid's info related components to Kid's folder in Dashboard
