+++ src/components/Dashboard/Kids/KidsInfo/KidForm.jsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { Col, Form, Row, Image, Container } from 'react-bootstrap';
 
 import { createKid } from '@api';
-import ModalKid from '@components/ModalKid/ModalKid';
+import ModalKid from '@components/Dashboard/Kids/ModalKid/ModalKid';
 import { CustomButton } from '@components/shared/ui/Button/CustomButton';
 import edit from '@media/icons/edit.svg';
