+++ src/components/Dashboard/Kids/KidsInfo/constants/allergies.js
